Very sloppy addition.
This 100% has a race condition in it... I think? I'd like to have some review done for it. The main event is that the message is daisy-chained through the clients until it reaches the original sender (or same string received). There literally has to be a better way to do this, but this is (in theory) what it should look like given a better solution.