File logging + SIGINT handled!
The issue is the `cargo run` command. It has it's own signal handler that terminates the main thread. For the signal handler to properly work, (Wait 10 seconds and then shut down server), you must run a built version of the program. Logging is added, file name is NOT dynamic upon program start, however, and filepath is next to source files. This can be changed, although I'm not sure where either of you would want it.