Command line args + shutdown/startup to history.log
Added command line args to both client and server. Also added message to be sent to history.log upon server startup and server shutdown. Pushing to main as it is a very simple addition. Changed exit codes to 0. Reason is because, while the program does receive an error, and I do believe in this instance it would be correct to exit with value of 0, I think it could be interpreted by Schwes as 'program crashes' accidentally. Simply print error message and exit instead.